# blog/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Blog",
        "contact": {},
        "version": "2.0"
    },
    "paths": {
        "/article/{id}": {
            "get": {
                "description": "Return article for given ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "External"
                ],
                "summary": "Return article for given ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.SuccessResponseModel"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseError"
                        }
                    }
                }
            }
        },
        "/articles": {
            "get": {
                "description": "Return list of articles",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "External"
                ],
                "summary": "Return list of articles",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.SuccessResponseModel"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseError"
                        }
                    }
                }
            },
            "post": {
                "description": "Create New Article",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "External"
                ],
                "summary": "Create Article",
                "parameters": [
                    {
                        "description": "Article",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Article"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.SuccessResponseModel"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Article": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "model.ResponseError": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "data"
                },
                "message": {
                    "description": "Error message",
                    "type": "string"
                },
                "status": {
                    "description": "Http status code",
                    "type": "integer"
                }
            }
        },
        "model.SuccessResponseModel": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "data"
                },
                "message": {
                    "description": "Error message",
                    "type": "string"
                },
                "status": {
                    "description": "Http status code",
                    "type": "integer"
                }
            }
        }
    }
}