# blog/internal/model/article.go
package model

// Article model contains all required data for articles
type Article struct {
	ID      int    `json:"id"`
	Title   string `json:"title"`
	Content string `json:"content"`
	Author  string `json:"author"`
}